SO 4313-7847: Separate polynomial.[ch] from pm43.c test code
